+++ src/App/Pages/PatientDetails/index.jsx
@@ -1,95 +1,22 @@
 import { useState, useEffect } from "react";
-import { useLocation, useParams } from "react-router-dom";
-import { DatePicker, Typography, Spin } from "antd";
+import { useLocation } from "react-router-dom";
+import { DatePicker, Typography } from "antd";
 import dayjs from "dayjs";
-
-import {
-  getNeurologyData,
-  getLabsData,
-  getVentilationData,
-} from "../../../services";
-import NeurologyTable from "../NeurologyTable";
-import VentilationTable from "../VentilationTable";
-import TabContent from "../../Table";
-import withDataTabs from "../WithTabsData";
-import { useQuery } from "react-query";
 import { parseQueryString } from "../../../Utils/functions";
-import LabsTable from "../LabsTable";
+import withDataTabs from "../../../components/WithTabsData";
+import TabContent from "../../../components/TabContent";
+import { TABS_CONFIG } from "./Patient.config";
 
 const { Title } = Typography;
-
-// Define sub-tab configurations
-const NEUROLOGY_SUBTABS = [
-  { title: "GCS", filterKey: "GCS", Component: NeurologyTable },
-  { title: "Pupil", filterKey: "Pupil", Component: NeurologyTable },
-  { title: "Strength", filterKey: "Strength", Component: NeurologyTable },
-  { title: "Orientation", filterKey: "Orientation", Component: NeurologyTable },
-  { title: "Motor", filterKey: "Motor", Component: NeurologyTable },
-];
-
-const VENTILATION_SUBTABS = [
-  { title: "Setting", filterKey: "Setting", Component: VentilationTable },
-  {
-    title: "Observation",
-    filterKey: "Observation",
-    Component: VentilationTable,
-  },
-];
-
-// Main tabs configuration
-const TABS_CONFIG = (stayId) => [
-  {
-    title: "Neurology",
-    key: "neurology",
-    fetchData: (date, type) => getNeurologyData(stayId, date, type),
-    subTabs: NEUROLOGY_SUBTABS,
-  },
-  {
-    title: "Labs",
-    key: "labs",
-    fetchData: (date) => getLabsData(stayId, date),
-    subTabs: null,
-    Component: LabsTable,
-  },
-  {
-    title: "Ventilation",
-    key: "ventilation",
-    fetchData: (date) => getVentilationData(date),
-    subTabs: VENTILATION_SUBTABS,
-  },
-];
 
 const PatientDetails = () => {
   const restParams = useLocation();
   const { stayId, lastDate } = parseQueryString(restParams?.search);
-
   const [selectedDate, setSelectedDate] = useState(dayjs());
 
   useEffect(() => {
     setSelectedDate(dayjs(lastDate));
   }, [lastDate]);
-
-  // Fetch neurology data to determine the last available date
-  const { data: neurologyData, isLoading: loadingNeurology } = useQuery(
-    ["neurology", stayId, dayjs(lastDate).isValid() ? lastDate : null],
-    () =>
-      getNeurologyData(
-        stayId,
-        dayjs(lastDate).isValid() ? lastDate : null,
-        "GCS"
-      ),
-    {
-      onSuccess: (data) => {
-        // Set default date to last available date in neurology data
-        // if (data && data.length > 0) {
-        //   const lastDatee = dayjs(data[data.length - 1].date).format(
-        //     "YYYY-MM-DD"
-        //   );
-        //   setSelectedDate(lastDatee);
-        // }
-      },
-    }
-  );
 
   const WrappedTabs = withDataTabs(TabContent);
 
@@ -98,20 +25,19 @@ const PatientDetails = () => {
       <Title level={2}>Patient Details: Stay ID {stayId}</Title>
       <DatePicker
         value={selectedDate}
-        onChange={setSelectedDate}
+        onChange={(date) => {
+          // Todo: Logic to change date with in query params
+          setSelectedDate(date);
+        }}
         format="YYYY-MM-DD"
         style={{ marginBottom: "20px" }}
       />
-      {loadingNeurology ? (
-        <Spin />
-      ) : (
-        <WrappedTabs
-          tabsConfig={TABS_CONFIG(stayId)}
-          date={selectedDate}
-          setSelectedDate={setSelectedDate}
-          stayId={stayId} // Pass stayId
-        />
-      )}
+      <WrappedTabs
+        tabsConfig={TABS_CONFIG(stayId)}
+        date={selectedDate}
+        setSelectedDate={setSelectedDate}
+        stayId={stayId}
+      />
     </div>
   );
 };
