+++ src/App/index.jsx
@@ -1,9 +1,6 @@
-import { Navigate, Route, Routes, BrowserRouter } from "react-router-dom";
+import { Route, Routes, BrowserRouter } from "react-router-dom";
 import LandingPage from "./Pages/LandingPage";
 import PatientDetails from "./Pages/PatientDetails";
-import Neurology from "./Pages/NeurologyTable";
-import Labs from "./Pages/LabsTable";
-import Ventilation from "./Pages/VentilationTable";
 
 function App() {
   return (
@@ -11,10 +8,10 @@ function App() {
       <Routes>
         <Route path="/" element={<LandingPage />} />
         <Route path="patients" element={<PatientDetails />}>
-          <Route index element={<Navigate to="neurology" replace />} />
-          <Route path="neurology" element={<Neurology />} />
+          {/* <Route index element={<Navigate to="neurology" replace />} /> */}
+          {/* <Route path="neurology" element={<Neurology />} />
           <Route path="labs" element={<Labs />} />
-          <Route path="ventilation" element={<Ventilation />} />
+          <Route path="ventilation" element={<Ventilation />} /> */}
         </Route>
       </Routes>
     </BrowserRouter>
