+++ src/components/LabsTable/LabsTable.mock.js
@@ -1,0 +1,9 @@
+export const columns = [
+  { title: "Subject ID", dataIndex: "subjectId", key: "subjectId" },
+  { title: "Stay ID", dataIndex: "stayId", key: "stayId" },
+  { title: "Lab Test", dataIndex: "testName", key: "testName" },
+  { title: "Result", dataIndex: "result", key: "result" },
+  { title: "Unit", dataIndex: "unit", key: "unit" },
+  { title: "Date", dataIndex: "date", key: "date" },
+  { title: "Time", dataIndex: "time", key: "time" },
+];
