+++ src/services/index.js
@@ -1,1 +1,1 @@
-export * from "./getAllStays";
+export * from "./getApiData";
